pyproject.toml
[project]
name = "ale"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matrix-nio>=0.25.2",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
    "ty>=0.0.16",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # pep8-naming
    "N",
]
